Move mvsData/Rgb.hpp to image module
