+++ src/aliceVision/depthMap/cuda/PlaneSweepingCuda.hpp
@@ -7,12 +7,12 @@
 #pragma once
 
 #include <aliceVision/image/Color.hpp>
+#include <aliceVision/image/Rgb.hpp>
 #include <aliceVision/mvsData/geometry.hpp>
 #include <aliceVision/mvsData/Pixel.hpp>
 #include <aliceVision/mvsData/Point2d.hpp>
 #include <aliceVision/mvsData/Point3d.hpp>
 #include <aliceVision/mvsData/Point4d.hpp>
-#include <aliceVision/mvsData/Rgb.hpp>
 #include <aliceVision/mvsData/StaticVector.hpp>
 #include <aliceVision/mvsData/Voxel.hpp>
 #include <aliceVision/mvsUtils/ImagesCache.hpp>
