+++ src/aliceVision/fuseCut/DelaunayGraphCut.hpp
@@ -7,8 +7,8 @@
 #pragma once
 
 #include <aliceVision/system/Logger.hpp>
+#include <aliceVision/image/Rgb.hpp>
 #include <aliceVision/mvsData/Point3d.hpp>
-#include <aliceVision/mvsData/Rgb.hpp>
 #include <aliceVision/mvsData/StaticVector.hpp>
 #include <aliceVision/mvsData/Voxel.hpp>
 #include <aliceVision/mvsUtils/common.hpp>
