+++ src/aliceVision/image/Rgb.hpp
