+++ src/aliceVision/image/all.hpp
@@ -24,6 +24,7 @@
 #include "aliceVision/image/io.hpp"
 #include "aliceVision/image/convolutionBase.hpp"
 #include "aliceVision/image/convolution.hpp"
+#include "aliceVision/image/Rgb.hpp"
 #include "aliceVision/image/Sampler.hpp"
 #include "aliceVision/image/convertionOpenCV.hpp"
