+++ src/aliceVision/mvsData/imageIO.cpp
@@ -9,7 +9,7 @@
 #include <aliceVision/system/Logger.hpp>
 #include <aliceVision/image/Color.hpp>
 #include <aliceVision/image/LegacyImage.hpp>
-#include <aliceVision/mvsData/Rgb.hpp>
+#include <aliceVision/image/Rgb.hpp>
 #include <aliceVision/image/imageAlgo.hpp>
 
 #include <OpenImageIO/imageio.h>
